[PM-2966] Use crypto primitives from mantis-crypto
