+++ metronome/networking/test/src/io/iohk/metronome/networking/RemoteConnectionManagerTestUtils.scala
@@ -65,10 +65,13 @@ object RemoteConnectionManagerTestUtils {
 
   object NodeInfo {
     def generateRandom(secureRandom: SecureRandom): NodeInfo = {
-      val keyPair = CryptoUtils.generateSecp256k1KeyPair(secureRandom)
+      val keyPair =
+        metronome.crypto.Secp256k1Utils.generateSecp256k1KeyPair(secureRandom)
       NodeInfo(
         keyPair,
-        Secp256k1Key(CryptoUtils.secp256k1KeyPairToNodeId(keyPair))
+        Secp256k1Key(
+          metronome.crypto.Secp256k1Utils.secp256k1KeyPairToNodeId(keyPair)
+        )
       )
     }
   }
